requirements.txt
# Core
fastapi>=0.128.0
uvicorn>=0.39.0
pydantic>=2.12.0
python-dotenv>=1.0.0

# Data & plotting
pandas>=2.0.0
numpy>=2.0.0
matplotlib>=3.9.0

# LLM / Agent
langchain-google-genai>=2.1.0
langchain-core>=0.3.0
langgraph>=0.6.0
google-api-core>=2.0.0

# Dev / test
pytest>=8.0.0
httpx>=0.27.0
